refactor: BackupPayload
